revert migration to util.ts
