+++ lib/manager/regex/index.ts
@@ -1,21 +1,92 @@
+import { URL } from 'url';
+import { logger } from '../../logger';
 import { regEx } from '../../util/regex';
+import * as template from '../../util/template';
 import type {
   CustomExtractConfig,
   PackageDependency,
   PackageFile,
   Result,
 } from '../types';
-import {
-  createDependency,
-  mergeExtractionTemplate,
-  mergeGroups,
-  regexMatchAll,
-  validMatchFields,
-} from './util';
 
 export const defaultConfig = {
   pinDigests: false,
 };
+
+const validMatchFields = [
+  'depName',
+  'lookupName',
+  'currentValue',
+  'currentDigest',
+  'datasource',
+  'versioning',
+  'extractVersion',
+  'registryUrl',
+  'depType',
+];
+
+interface ExtractionTemplate {
+  groups: Record<string, string>;
+  replaceString: string;
+}
+
+function regexMatchAll(regex: RegExp, content: string): RegExpMatchArray[] {
+  const matches: RegExpMatchArray[] = [];
+  let matchResult;
+  do {
+    matchResult = regex.exec(content);
+    if (matchResult) {
+      matches.push(matchResult);
+    }
+  } while (matchResult);
+  return matches;
+}
+
+function createDependency(
+  extractionTemplate: ExtractionTemplate,
+  config: CustomExtractConfig,
+  dep?: PackageDependency
+): PackageDependency {
+  const dependency = dep || {};
+  const { groups, replaceString } = extractionTemplate;
+
+  function updateDependency(field: string, value: string): void {
+    switch (field) {
+      case 'registryUrl':
+        // check if URL is valid and pack inside an array
+        try {
+          const url = new URL(value).toString();
+          dependency.registryUrls = [url];
+        } catch (err) {
+          logger.warn({ value }, 'Invalid regex manager registryUrl');
+        }
+        break;
+      default:
+        dependency[field] = value;
+        break;
+    }
+  }
+
+  for (const field of validMatchFields) {
+    const fieldTemplate = `${field}Template`;
+    if (config[fieldTemplate]) {
+      try {
+        const compiled = template.compile(config[fieldTemplate], groups, false);
+        updateDependency(field, compiled);
+      } catch (err) {
+        logger.warn(
+          { template: config[fieldTemplate] },
+          'Error compiling template for custom manager'
+        );
+        return null;
+      }
+    } else if (groups[field]) {
+      updateDependency(field, groups[field]);
+    }
+  }
+  dependency.replaceString = replaceString;
+  return dependency;
+}
 
 function handleAny(
   content: string,
@@ -31,6 +102,23 @@ function handleAny(
         config
       )
     );
+}
+
+function mergeGroups(
+  mergedGroup: Record<string, string>,
+  secondGroup: Record<string, string>
+): Record<string, string> {
+  return { ...mergedGroup, ...secondGroup };
+}
+
+export function mergeExtractionTemplate(
+  base: ExtractionTemplate,
+  addition: ExtractionTemplate
+): ExtractionTemplate {
+  return {
+    groups: mergeGroups(base.groups, addition.groups),
+    replaceString: addition.replaceString ?? base.replaceString,
+  };
 }
 
 function handleCombination(
